Heavily refactored and compartmentalized mavlink code; tested and everything still works
