+++ include/mavlink.h
@@ -1,15 +1,15 @@
 #pragma once
 
-#include <stdint.h>
+#include <mavlink/v1.0/mavlink_types.h>
+
+#define MAVLINK_USE_CONVENIENCE_FUNCTIONS
+extern mavlink_system_t mavlink_system;
+void comm_send_ch(mavlink_channel_t chan, uint8_t ch);
+
+#include <mavlink/v1.0/rosflight/mavlink.h>
 
 // function declarations
 void init_mavlink(void);
 
-void mavlink_receive(void);
-
-void start_send_param_list(void);
-
 void send_heartbeat(void);
 void send_imu(int16_t ax, int16_t ay, int16_t az, int16_t gx, int16_t gy, int16_t gz);
-
-void mavlink_send_low_priority(void);
