--- include/mavlink_bridge.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#include <mavlink/v1.0/mavlink_types.h>
-extern mavlink_system_t mavlink_system;
-
-#define MAVLINK_USE_CONVENIENCE_FUNCTIONS
-#include <mavlink/v1.0/rosflight/mavlink.h>
