+++ include/mavlink_receive.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void mavlink_receive(void);
