+++ include/param.h
@@ -3,7 +3,7 @@
 #include <stdbool.h>
 #include <stdint.h>
 
-#include "mavlink_bridge.h"
+#include "mavlink.h"
 
 #define PARAMS_COUNT 3
 #define PARAMS_NAME_LENGTH MAVLINK_MSG_PARAM_SET_FIELD_PARAM_ID_LEN
