--- src/mavlink_bridge.c
@@ -1,16 +1,0 @@
-#include <stdint.h>
-
-#include <breezystm32/breezystm32.h>
-#include <breezystm32/drv_serial.h>
-
-#include "mavlink_bridge.h"
-
-mavlink_system_t mavlink_system;
-
-inline void comm_send_ch(mavlink_channel_t chan, uint8_t ch)
-{
-  if (chan == MAVLINK_COMM_0)
-  {
-    serialWrite(Serial1, ch);
-  }
-}
